estimates now changing, but really small kalman gains.
something seems off with the update step, plugged it into gpt and said that yaw couldnt be calculated with just gyroscope accelerometer data. also not sure to trust gyroscope data as much as accelerometer.

tried comparing to onb-errorekfraa to see how it was implemented. seemed similar but a little confused on whats going on there.